chore: Refactoring ColumnOrdering into separate base, types, and utils
