+++ packages/table-core/src/core/table.ts
@@ -26,7 +26,7 @@ import { Headers } from './headers'
 import { ColumnFaceting } from '../features/column-faceting/ColumnFaceting'
 import { ColumnFiltering } from '../features/column-filtering/ColumnFiltering'
 import { ColumnGrouping } from '../features/column-grouping/ColumnGrouping'
-import { ColumnOrdering } from '../features/ColumnOrdering'
+import { ColumnOrdering } from '../features/column-ordering/ColumnOrdering'
 import { ColumnPinning } from '../features/ColumnPinning'
 import { ColumnSizing } from '../features/ColumnSizing'
 import { ColumnVisibility } from '../features/column-visibility/ColumnVisibility'
